Move satisfiesconstraints trait into `source_code_unit.rs`
